requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0
pytest>=7.0.0
pytest-mock>=3.8.0
pytest-cov>=4.0.0